dtorrtchqe/Twilio_call_transcriber/requirements.txt
pyngrok ==7.2.0
ibm_watson 
flask_socketio 
ngrok 
requests 
twilio 
pytz 
python-dotenv 
PyJWT 
pyaudio 
websocket-client 
flask_sock
numpy
flask